+++ src/Terrabuild.Scripting/Contracts.fs
@@ -1,6 +1,5 @@
 namespace Terrabuild
 
-open System
 open Collections
 
 module ScriptingContracts =
@@ -78,21 +77,3 @@ module ScriptingContracts =
         | Cache of Cacheability
 
     type ScriptDescriptor = Map<string, ExportFlag list>
-
-    [<AbstractClass>]
-    [<AttributeUsage(AttributeTargets.Method, AllowMultiple = false)>]
-    type CacheableAttribute(cacheability: Cacheability) =
-        inherit Attribute()
-        member _.Cacheability = cacheability
-
-    type ExternalCacheAttribute() =
-        inherit CacheableAttribute(Cacheability.External)
-
-    type RemoteCacheAttribute() =
-        inherit CacheableAttribute(Cacheability.Remote)
-
-    type LocalCacheAttribute() =
-        inherit CacheableAttribute(Cacheability.Local)
-
-    type NoCacheAttribute() =
-        inherit CacheableAttribute(Cacheability.Never)